[dynamo][guards] Flush cache to more accurately measure guard overhead
We observed that guard overhead at runtime using profiler traces was
higher than reported in this profiling function at the compile time.
After investigation, we found that f_locals are already in cache and
that was causing the guard overhead to be way smaller while profiling
during the compilation. To be more realistic, we flush the cache here.

Profiling the guard overhead during compilation (in addition to at
runtime) allows faster iteration time, and logging in tlparse and
internal databases.

ghstack-source-id: 4ad5bbcc24955f2fc237a3b2474e7a0bb8c432e2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/154764